Moved polygon utilities to its own file
